renamed recommended folder to Likes, updated imports
