+++ src/pages/profile/UserProfile.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom';
 import { Grommet, Tab, Tabs } from 'grommet';
 import styled from 'styled-components';
 
-import Likes from '../../components/recommended/Likes';
+import Likes from '../../components/likes/Likes';
 import { customWrapper } from '../../components/mixins';
 import SidebarById from '../../components/sidebar/SidebarById';
 import ProfileById from './ProfileById';
